sabzee_v3/Sabzee Server/plant_disease_api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
torch>=2.2.0
torchvision>=0.17.0
Pillow>=9.0.0
requests==2.26.0 